requirements.txt
openai-whisper
pydub
scipy
numpy
matplotlib